+++ src/hooks/queries/useArticlesQuery.ts
@@ -1,11 +1,23 @@
 import { useCallback } from 'react'
 import { useQuery } from 'react-query'
+import { gql } from 'graphql-request'
 import { Article } from '../../types/dataTypes'
-import { GET_ARTICLES } from '../../queries/queries'
 import QUERY_KEY from '../../constants/query-key'
 import useStore from '../../store/useStore'
 
-interface ArticlesRes {
+const GET_ARTICLES = gql`
+  query GetArticles {
+    articles(order_by: { updated_at: desc }) {
+      id
+      created_at
+      updated_at
+      title
+      content
+    }
+  }
+`
+
+interface QueryResult {
   articles: Article[]
 }
 
@@ -13,7 +25,7 @@ export const useArticlesQuery = () => {
   const graphQLClient = useStore((state) => state.graphQLClient)
 
   const fetchArticles = useCallback(async () => {
-    return (await graphQLClient.request<ArticlesRes>(GET_ARTICLES)).articles
+    return (await graphQLClient.request<QueryResult>(GET_ARTICLES)).articles
   }, [graphQLClient])
 
   return useQuery<Article[], Error>({
