+++ src/hooks/useFetchCurrentUser.ts
@@ -1,9 +1,18 @@
 import { useCallback } from 'react'
+import { gql } from 'graphql-request'
 import { User } from '../types/dataTypes'
-import { GET_CURRENT_USER } from '../queries/queries'
 import useStore from '../store/useStore'
 
-interface CurrentUserRes {
+const GET_CURRENT_USER = gql`
+  query GetCurrentUser($id: String!) {
+    users_by_pk(id: $id) {
+      id
+      email
+    }
+  }
+`
+
+interface QueryResult {
   users_by_pk: User
 }
 
@@ -12,7 +21,7 @@ export const useFetchCurrentUser = () => {
 
   const fetchCurrentUser = useCallback(
     async (firebaseUid: string) => {
-      const { users_by_pk: data } = await graphQLClient.request<CurrentUserRes>(GET_CURRENT_USER, {
+      const { users_by_pk: data } = await graphQLClient.request<QueryResult>(GET_CURRENT_USER, {
         id: firebaseUid,
       })
       return data
